refactor: 레포지토리를 인터페이스로 생성
